move tipeee and twitch
