+++ bot/main/talium/TwitchBot.java
@@ -3,11 +3,11 @@ package talium;
 import jakarta.annotation.PreDestroy;
 import jakarta.persistence.PreRemove;
 import org.springframework.context.ConfigurableApplicationContext;
-import talium.inputs.TipeeeStream.DonationRepo;
-import talium.inputs.TipeeeStream.TipeeeConfig;
-import talium.inputs.TipeeeStream.TipeeeInput;
-import talium.inputs.Twitch4J.Twitch4JInput;
-import talium.inputs.Twitch4J.TwitchConfig;
+import talium.TipeeeStream.DonationRepo;
+import talium.TipeeeStream.TipeeeConfig;
+import talium.TipeeeStream.TipeeeInput;
+import talium.Twitch4J.Twitch4JInput;
+import talium.Twitch4J.TwitchConfig;
 import talium.oauthConnector.OauthAccountRepo;
 import talium.system.StopWatch;
 import talium.system.coinsWatchtime.WIPWatchtimeCommandServer;
