+++ bot/main/talium/system/Out.java
@@ -3,8 +3,8 @@ package talium.system;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
-import talium.inputs.Twitch4J.TwitchAPINoop;
-import talium.inputs.Twitch4J.TwitchApi;
+import talium.Twitch4J.TwitchAPINoop;
+import talium.Twitch4J.TwitchApi;
 import talium.system.stringTemplates.Template;
 import talium.system.stringTemplates.TemplateService;
 import talium.system.templateParser.exeptions.*;
