+++ bot/main/talium/system/twitchCommands/cooldown/CooldownService.java
@@ -6,8 +6,8 @@ import org.checkerframework.dataflow.qual.Impure;
 import org.checkerframework.dataflow.qual.Pure;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import talium.inputs.Twitch4J.ChatMessage;
-import talium.inputs.Twitch4J.TwitchUser;
+import talium.Twitch4J.ChatMessage;
+import talium.Twitch4J.TwitchUser;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
